Fixed a bug when capturing external process-output
All data are processed the then close-event gets triggered - the exit-event on
the other hand does not guarantee (although it can happen) that all output
data are already delivered yet. So, I've fixed that.
Also, it is considered to be an antipattern to accumulate the process-output
as V8 in-memory strings. It's better to work with buffers as long as possible to
keep the string-munging out of V8's memory