Fixed a bug in the font-post-processor
The processor made use of xmldom which, as it turned out, changes attribute values
when they contain the &gt; escaping-sequence. This on the other hand leads to an issue
when trying to parse the output of the post-processing step. Therefore we had to
provide a nasty, but working solution to the problem without xml-parsing stuff but with
a simple string-manipulation.