Now the link post-processor is also contained.
The post-processor now enhances all contained links with an onclick
handler. The function (which is also injected) will check if a companion
script (more precisely a function called 'wuml.navigateTo') is present
and will delegate the click-event to that function. If there is no companion
script available nothing special happens and the link is executed.